Convert tool projects to file-based apps
Replace the five standalone tool projects (DbContextDataMigrator,
DbMigrationApplier, ImageSizeCalculator, Tools.ImageVariantUpdater,
Tools.ShareInheritableMigrator) with file-based .NET apps at the src/
root, following the existing SampleDataSeeder.cs and VideoSizeBackfill.cs
pattern. Each new file declares its SDK/project references via #:sdk,
#:property PublishAot=false, and #:project directives.

The five .csproj folders are removed and their entries, configuration
blocks, NestedProjects mappings, and the now-empty Tools solution folder
are stripped from Recollections.sln.
