--- src/Recollections.DbContextDataMigrator/Recollections.DbContextDataMigrator.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>Neptuo.Recollections</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Recollections.Accounts.Data\Recollections.Accounts.Data.csproj" />
-    <ProjectReference Include="..\Recollections.Entries.Data\Recollections.Entries.Data.csproj" />
-  </ItemGroup>
-
-</Project>
