--- src/Recollections.DbMigrationApplier/Recollections.DbMigrationApplier.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>Neptuo.Recollections</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Recollections.Accounts.Data\Recollections.Accounts.Data.csproj" />
-    <ProjectReference Include="..\Recollections.Data.Ef\Recollections.Data.Ef.csproj" />
-    <ProjectReference Include="..\Recollections.Entries.Data\Recollections.Entries.Data.csproj" />
-  </ItemGroup>
-
-</Project>
